Move notification components to new folder structure
